refactor : Lotto 결과 생성 정적 팩토리 메서드로 분리
