+++ src/main/java/lotto/domain/Lottos.java
@@ -11,15 +11,6 @@ public class Lottos {
         this.lottos = lottos;
     }
 
-    public LottoResult createResult(final WinLotto winLotto) {
-        final Map<Rank, Integer> resultMap = Rank.initResultMap();
-        for (Lotto lotto : lottos) {
-            Rank rank = winLotto.matchResult(lotto);
-            resultMap.replace(rank, resultMap.get(rank) + 1);
-        }
-        return new LottoResult(resultMap);
-    }
-
     public List<Lotto> getLottos() {
         return List.copyOf(lottos);
     }
